+++ samples/sample-express-mongodb/src/controllers/paging.controller.ts
@@ -1,20 +1,22 @@
-import {Controller, CursorRequest, Get, PagingRequest, QueryParams} from "@nodeboot/core";
+import {Controller, CursorPage, CursorRequest, Get, Page, PagingRequest, QueryParams} from "@nodeboot/core";
 import {UserModel} from "../models";
 import {Logger} from "winston";
 import {PagingUserRepository} from "../persistence";
-import {CursorPage, Page} from "@nodeboot/starter-persistence";
+import {ResponseSchema} from "@nodeboot/starter-openapi";
 
 @Controller("/paging", "v1")
 export class PagingUserController {
     constructor(private readonly userRepository: PagingUserRepository, private readonly logger: Logger) {}
 
     @Get("/paginated")
+    @ResponseSchema(Page<UserModel>)
     async getUsersPaginated(@QueryParams() paging: PagingRequest): Promise<Page<UserModel>> {
         this.logger.info(`Getting paginated users`);
         return this.userRepository.findPaginated({}, paging);
     }
 
     @Get("/cursor")
+    @ResponseSchema(CursorPage<UserModel>)
     async getUsersCursorPaginated(@QueryParams() cursorRequest: CursorRequest): Promise<CursorPage<UserModel>> {
         this.logger.info(`Getting cursor paginated users`);
         return this.userRepository.findCursorPaginated({}, cursorRequest);