+++ starters/persistence/src/repository/MongoPagingAndSortingRepository.ts
@@ -1,6 +1,6 @@
 import {FindOptionsWhere, MongoRepository, ObjectLiteral} from "typeorm";
-import {CursorPage, CursorRequest, Page, PagingRequest} from "./types";
 import {SortOrder} from "@nodeboot/core";
+import {CursorPage, CursorRequest, Page, PagingRequest} from "@nodeboot/core/src";
 
 /**
  * A generic MongoDB repository that provides both offset-based and cursor-based pagination.