+++ starters/persistence/src/repository/index.ts
@@ -1,3 +1,2 @@
 export {PagingAndSortingRepository} from "./PagingAndSortingRepository";
 export {MongoPagingAndSortingRepository} from "./MongoPagingAndSortingRepository";
-export * from "./types";