+++ packages/core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @nodeboot/core
 
+## 1.5.0
+
+### Minor Changes
+
+-   Cleanup OpenAPI decorators like Model and Property to improve type resolution
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.4.0
+    -   @nodeboot/config@1.1.5
+    -   @nodeboot/di@1.1.5
+    -   @nodeboot/engine@1.1.5
+
 ## 1.4.1
 
 ### Patch Changes