+++ servers/express-server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @nodeboot/express-server
 
+## 1.2.4
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.4.0
+    -   @nodeboot/core@1.5.0
+    -   @nodeboot/engine@1.1.5
+
 ## 1.2.3
 
 ### Patch Changes