+++ starters/actuator/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @nodeboot/starter-actuator
 
+## 1.1.7
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.4.0
+    -   @nodeboot/config@1.1.5
+    -   @nodeboot/engine@1.1.5
+
 ## 1.1.6
 
 ### Patch Changes