+++ starters/firebase/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @nodeboot/starter-firebase
 
+## 1.0.7
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.4.0
+    -   @nodeboot/core@1.5.0
+    -   @nodeboot/config@1.1.5
+    -   @nodeboot/engine@1.1.5
+
 ## 1.0.6
 
 ### Patch Changes