+++ starters/http/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nodeboot/starter-http
 
+## 2.0.3
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.4.0
+
 ## 2.0.2
 
 ### Patch Changes