+++ starters/scheduler/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nodeboot/starter-scheduler
 
+## 1.0.5
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.4.0
+
 ## 1.0.4
 
 ### Patch Changes