+++ tests/jest-express-demo/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @nodeboot/jest-express-demo
 
+## 1.0.8
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/starter-persistence@1.6.0
+    -   @nodeboot/context@1.4.0
+    -   @nodeboot/core@1.5.0
+    -   @nodeboot/config@1.1.5
+    -   @nodeboot/di@1.1.5
+    -   @nodeboot/engine@1.1.5
+    -   @nodeboot/express-server@1.2.4
+
 ## 1.0.7
 
 ### Patch Changes